requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
passlib
email-validator
reportlab
python-multipart
requests
